Fixed read-isExactly and conditional item.links check

+++ UpdateHandler.class.ts
@@ -44,7 +44,7 @@ export abstract class UpdateHandler extends LambdaHandler {
 
 
         private unlinkOldRecords() {
-          if (!this.request.item.links._unlink) return
+          if (!this.request.item.links || !this.request.item.links._unlink) return
           for (let [ table, ids ] of Object.entries(this.request.item.links._unlink as any)) {
             Object.keys(ids).forEach(id => this.unlink({ table: table, id: id }))
           }
@@ -80,7 +80,7 @@ export abstract class UpdateHandler extends LambdaHandler {
 
 
         private linkNewRecords() {
-          if (!this.request.item.links._link) return
+          if (!this.request.item.links || !this.request.item.links._link) return
           for (let [ table, ids ] of Object.entries(this.request.item.links._link as any)) {
             Object.keys(ids).forEach(id => this.link({ table: table, id: id }))
           }
